Intial pass at type annotating slackbridge
This makes mypy checking stricter (and also moves to using a config
file). The only question I faced when working on this was how to
represent the JSON responses of the Slack API. I decided on `Dict[str,
Any]`, which should be sufficient for now. Later if interested it would
be a fun project to convert these to
[TypedDict](https://mypy.readthedocs.io/en/latest/more_types.html#typeddict),
which would give us a bit more type safety. I will open an issue on that
once this is merged (and would happy to mentor that).